changes from plan 9

+++ src/libdraw/md-defont.c
@@ -21,7 +21,7 @@ getmemdefont(void)
 	 * declared as char*, not u32int*.
 	 */
 	p = (char*)defontdata;
-	n = (ulong)p & 3;
+	n = (uintptr)p & 3;
 	if(n != 0){
 		memmove(p+(4-n), p, sizeofdefont-n);
 		p += 4-n;
